requirements.txt
fastapi>=0.115,<1
uvicorn>=0.30,<1
requests>=2.32,<3
python-dotenv>=1,<2
langchain>=0.2,<0.4
langchain-openai>=0.1,<0.4